feat!: by default determine version from Git tags
The plugin now is not configurable any more.
However, there is a plugin variant that can be used as before, where the
configuration can be adapted and also a version file used if desired.
