+++ src/main/groovy/to/wetransform/gradle/version/ConfigurableVersionPlugin.groovy
@@ -1,0 +1,10 @@
+package to.wetransform.gradle.version
+
+/**
+ * Version plugin that allows customization through VersionExtension.
+ */
+class ConfigurableVersionPlugin extends AbstractVersionPlugin {
+  ConfigurableVersionPlugin() {
+    super(true)
+  }
+}
